feat: update code util
