+++ packages/simm-fetch/src/utils/handle-retry.ts
@@ -1,9 +1,28 @@
-import { AxiosResponse } from "axios";
 import { APIClientConfig } from "../types";
-export const handleRetry = async (
-  requestFn: () => Promise<AxiosResponse>,
-  config: APIClientConfig,
-): Promise<AxiosResponse> => {
-  //Todo retry API with option/config
-  config.isRetry = false;
-};
+export async function handleRetry(error: any, config: APIClientConfig): Promise<any> {
+  const { retry } = config;
+  const retryConfig = {
+    retryCount: 0,
+    retryDelay: retry?.delay || 1000,
+    maxRetries: retry?.attempts || 3,
+  };
+
+  const retryRequest = async () => {
+    try {
+      retryConfig.retryCount++;
+
+      return await error.config.adapter.request({
+        ...error.config,
+        url: error.config.url,
+      });
+    } catch (retryError) {
+      if (retryConfig.retryCount < retryConfig.maxRetries) {
+        await new Promise(resolve => setTimeout(resolve, retryConfig.retryDelay));
+        return retryRequest();
+      }
+      throw retryError;
+    }
+  };
+
+  return retryRequest();
+}
